starting pkg orders

+++ kamoney_sdk_dtos/account_history.go
@@ -1,15 +1,15 @@
 package kamoney_sdk_dtos
 
 type GetAccountHistoryRequestParams struct {
-	Page  int    `json:"page"`
-	Date  int    `json:"date"`
+	Page  int64  `json:"page"`
+	Date  int64  `json:"date"`
 	Nonce string `json:"nonce"`
 }
 
 type GetAccountHistoryRequestResponse struct {
 	Common
 	Data []struct {
-		ID      int    `json:"id"`
+		ID      int64  `json:"id"`
 		Created string `json:"created"`
 		History string `json:"history"`
 	} `json:"data"`
