+++ kamoney_sdk_dtos/services_order.go
@@ -10,12 +10,12 @@ type ServicesOrderRequestResponse struct {
 			Picture      string   `json:"picture"`
 			Name         string   `json:"name"`
 			Maintenance  bool     `json:"maintenance"`
-			LimitMin     int      `json:"limit_min"`
-			LimitMax     int      `json:"limit_max"`
-			FeeThreshold int      `json:"fee_threshold"`
-			FeeAmount    int      `json:"fee_amount"`
+			LimitMin     int64    `json:"limit_min"`
+			LimitMax     int64    `json:"limit_max"`
+			FeeThreshold int64    `json:"fee_threshold"`
+			FeeAmount    int64    `json:"fee_amount"`
 			Automatic    bool     `json:"automatic"`
-			AutomaticMax int      `json:"automatic_max"`
+			AutomaticMax int64    `json:"automatic_max"`
 			Info         []string `json:"info"`
 		} `json:"payment_slips"`
 		DirectTransfers struct {
@@ -23,12 +23,12 @@ type ServicesOrderRequestResponse struct {
 			Picture      string   `json:"picture"`
 			Name         string   `json:"name"`
 			Maintenance  bool     `json:"maintenance"`
-			LimitMin     int      `json:"limit_min"`
-			LimitMax     int      `json:"limit_max"`
-			FeeThreshold int      `json:"fee_threshold"`
+			LimitMin     int64    `json:"limit_min"`
+			LimitMax     int64    `json:"limit_max"`
+			FeeThreshold int64    `json:"fee_threshold"`
 			FeeAmount    float64  `json:"fee_amount"`
 			Automatic    bool     `json:"automatic"`
-			AutomaticMax int      `json:"automatic_max"`
+			AutomaticMax int64    `json:"automatic_max"`
 			Info         []string `json:"info"`
 		} `json:"direct_transfers"`
 		Pix struct {
@@ -36,12 +36,12 @@ type ServicesOrderRequestResponse struct {
 			Picture      string   `json:"picture"`
 			Name         string   `json:"name"`
 			Maintenance  bool     `json:"maintenance"`
-			LimitMin     int      `json:"limit_min"`
-			LimitMax     int      `json:"limit_max"`
-			FeeThreshold int      `json:"fee_threshold"`
-			FeeAmount    int      `json:"fee_amount"`
+			LimitMin     int64    `json:"limit_min"`
+			LimitMax     int64    `json:"limit_max"`
+			FeeThreshold int64    `json:"fee_threshold"`
+			FeeAmount    int64    `json:"fee_amount"`
 			Automatic    bool     `json:"automatic"`
-			AutomaticMax int      `json:"automatic_max"`
+			AutomaticMax int64    `json:"automatic_max"`
 			Info         []string `json:"info"`
 		} `json:"pix"`
 		DigitalProducts struct {
@@ -49,12 +49,12 @@ type ServicesOrderRequestResponse struct {
 			Picture      string   `json:"picture"`
 			Name         string   `json:"name"`
 			Maintenance  bool     `json:"maintenance"`
-			LimitMin     int      `json:"limit_min"`
-			LimitMax     int      `json:"limit_max"`
-			FeeThreshold int      `json:"fee_threshold"`
-			FeeAmount    int      `json:"fee_amount"`
+			LimitMin     int64    `json:"limit_min"`
+			LimitMax     int64    `json:"limit_max"`
+			FeeThreshold int64    `json:"fee_threshold"`
+			FeeAmount    int64    `json:"fee_amount"`
 			Automatic    bool     `json:"automatic"`
-			AutomaticMax int      `json:"automatic_max"`
+			AutomaticMax int64    `json:"automatic_max"`
 			Info         []string `json:"info"`
 		} `json:"digital_products"`
 	} `json:"data"`
